inserting repos and main code

+++ src/Dtos/IssueTicketRequest.java
@@ -5,7 +5,7 @@ import Models.VehicleType;
 public class IssueTicketRequest {
 
     private String vehicleNo;
-    private long gateId;
+    private int gateId;
     private String ownerName;
     private VehicleType vehicleType;
 
@@ -17,11 +17,11 @@ public class IssueTicketRequest {
         this.vehicleNo = vehicleNo;
     }
 
-    public long getGateId() {
+    public int getGateId() {
         return gateId;
     }
 
-    public void setGateId(long gateId) {
+    public void setGateId(int gateId) {
         this.gateId = gateId;
     }
